refactor k8s root ca secret access
mostly moving things around to make subsequent PR's smaller
