+++ pkg/certificate/certificate.go
@@ -4,7 +4,10 @@ import (
 	"math/rand"
 	time "time"
 
+	"github.com/rs/zerolog/log"
+
 	"github.com/openservicemesh/osm/pkg/certificate/pem"
+	"github.com/openservicemesh/osm/pkg/errcode"
 )
 
 const (
@@ -57,4 +60,24 @@ func (c *Certificate) ShouldRotate() bool {
 	intNoise := rand.Intn(noiseSeconds) // #nosec G404
 	secondsNoise := time.Duration(intNoise) * time.Second
 	return time.Until(c.GetExpiration()) <= (RenewBeforeCertExpires + secondsNoise)
+}
+
+// NewFromPEM is a helper returning a *certificate.Certificate from the PEM components given.
+func NewFromPEM(pemCert pem.Certificate, pemKey pem.PrivateKey) (*Certificate, error) {
+	x509Cert, err := DecodePEMCertificate(pemCert)
+	if err != nil {
+		// TODO(#3962): metric might not be scraped before process restart resulting from this error
+		log.Error().Err(err).Str(errcode.Kind, errcode.GetErrCodeWithMetric(errcode.ErrDecodingPEMCert)).
+			Msg("Error converting PEM cert to x509 to obtain serial number")
+		return nil, err
+	}
+
+	return &Certificate{
+		CommonName:   CommonName(x509Cert.Subject.CommonName),
+		SerialNumber: SerialNumber(x509Cert.SerialNumber.String()),
+		CertChain:    pemCert,
+		IssuingCA:    pem.RootCertificate(pemCert),
+		PrivateKey:   pemKey,
+		Expiration:   x509Cert.NotAfter,
+	}, nil
 }
