+++ pkg/certificate/providers/config.go
@@ -7,22 +7,19 @@ import (
 	cmmeta "github.com/jetstack/cert-manager/pkg/apis/meta/v1"
 	cmversionedclient "github.com/jetstack/cert-manager/pkg/client/clientset/versioned"
 	"github.com/pkg/errors"
-	corev1 "k8s.io/api/core/v1"
-	apierrors "k8s.io/apimachinery/pkg/api/errors"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 	"k8s.io/client-go/kubernetes"
 	"k8s.io/client-go/rest"
 
 	"github.com/openservicemesh/osm/pkg/certificate"
+	"github.com/openservicemesh/osm/pkg/certificate/castorage/k8s"
 	"github.com/openservicemesh/osm/pkg/certificate/providers/certmanager"
 	"github.com/openservicemesh/osm/pkg/certificate/providers/tresor"
 	"github.com/openservicemesh/osm/pkg/certificate/providers/vault"
 	"github.com/openservicemesh/osm/pkg/configurator"
 	"github.com/openservicemesh/osm/pkg/constants"
 	"github.com/openservicemesh/osm/pkg/debugger"
-	"github.com/openservicemesh/osm/pkg/errcode"
 	"github.com/openservicemesh/osm/pkg/messaging"
-	"github.com/openservicemesh/osm/pkg/version"
 )
 
 const (
@@ -32,8 +29,8 @@ const (
 	rootCertOrganization = "Open Service Mesh"
 )
 
-// NewCertificateProvider returns a new certificate provider and associated config
-func NewCertificateProvider(kubeClient kubernetes.Interface, kubeConfig *rest.Config, cfg configurator.Configurator, providerKind Kind,
+// GenerateCertificateManager returns a new certificate manager and associated config
+func GenerateCertificateManager(kubeClient kubernetes.Interface, kubeConfig *rest.Config, cfg configurator.Configurator, providerKind Kind,
 	providerNamespace string, caBundleSecretName string, tresorOptions TresorOptions, vaultOptions VaultOptions,
 	certManagerOptions CertManagerOptions, msgBroker *messaging.Broker) (*certificate.Manager, debugger.CertificateManagerDebugger, *Config, error) {
 	config := &Config{
@@ -61,26 +58,6 @@ func NewCertificateProvider(kubeClient kubernetes.Interface, kubeConfig *rest.Co
 	}
 
 	return certManager, certDebugger, config, nil
-}
-
-// NewCertificateProviderConfig returns a new certificate provider config
-func NewCertificateProviderConfig(kubeClient kubernetes.Interface, kubeConfig *rest.Config, cfg configurator.Configurator, providerKind Kind,
-	providerNamespace string, caBundleSecretName string, tresorOptions TresorOptions, vaultOptions VaultOptions,
-	certManagerOptions CertManagerOptions, msgBroker *messaging.Broker) *Config {
-	return &Config{
-		kubeClient:         kubeClient,
-		kubeConfig:         kubeConfig,
-		cfg:                cfg,
-		providerKind:       providerKind,
-		providerNamespace:  providerNamespace,
-		caBundleSecretName: caBundleSecretName,
-
-		tresorOptions:      tresorOptions,
-		vaultOptions:       vaultOptions,
-		certManagerOptions: certManagerOptions,
-
-		msgBroker: msgBroker,
-	}
 }
 
 // Validate validates the certificate provider config
@@ -160,50 +137,6 @@ func (c *Config) GetCertificateManager() (*certificate.Manager, debugger.Certifi
 	}
 }
 
-// GetCertificateFromSecret is a helper function that ensures creation and synchronization of a certificate
-// using Kubernetes Secrets backend and API atomicity.
-func GetCertificateFromSecret(ns string, secretName string, cert *certificate.Certificate, kubeClient kubernetes.Interface) (*certificate.Certificate, error) {
-	// Attempt to create it in Kubernetes. When multiple agents attempt to create, only one of them will succeed.
-	// All others will get "AlreadyExists" error back.
-	secretData := map[string][]byte{
-		constants.KubernetesOpaqueSecretCAKey:             cert.GetCertificateChain(),
-		constants.KubernetesOpaqueSecretRootPrivateKeyKey: cert.GetPrivateKey(),
-	}
-
-	secret := &corev1.Secret{
-		ObjectMeta: metav1.ObjectMeta{
-			Name:      secretName,
-			Namespace: ns,
-			Labels: map[string]string{
-				constants.OSMAppNameLabelKey:    constants.OSMAppNameLabelValue,
-				constants.OSMAppVersionLabelKey: version.Version,
-			},
-		},
-		Data: secretData,
-	}
-
-	if _, err := kubeClient.CoreV1().Secrets(ns).Create(context.TODO(), secret, metav1.CreateOptions{}); err == nil {
-		log.Info().Msgf("Secret %s/%s created in kubernetes", ns, secretName)
-	} else if apierrors.IsAlreadyExists(err) {
-		log.Info().Msgf("Secret %s/%s already exists in kubernetes, loading.", ns, secretName)
-	} else {
-		// TODO(#3962): metric might not be scraped before process restart resulting from this error
-		log.Error().Err(err).Str(errcode.Kind, errcode.GetErrCodeWithMetric(errcode.ErrCreatingCertSecret)).
-			Msgf("Error creating/retrieving certificate secret %s/%s", ns, secretName)
-		return nil, err
-	}
-
-	// For simplicity, we will load the certificate for all of them, this way the instance which created it
-	// and the ones that didn't share the same code.
-	cert, err := GetCertFromKubernetes(ns, secretName, kubeClient)
-	if err != nil {
-		log.Error().Err(err).Msg("Failed to fetch certificate from Kubernetes")
-		return nil, err
-	}
-
-	return cert, nil
-}
-
 // getTresorOSMCertificateManager returns a certificate manager instance with Tresor as the certificate provider
 func (c *Config) getTresorOSMCertificateManager() (*certificate.Manager, debugger.CertificateManagerDebugger, error) {
 	var err error
@@ -228,9 +161,13 @@ func (c *Config) getTresorOSMCertificateManager() (*certificate.Manager, debugge
 		return nil, nil, errors.Errorf("Root cert does not have a private key")
 	}
 
-	rootCert, err = GetCertificateFromSecret(c.providerNamespace, c.caBundleSecretName, rootCert, c.kubeClient)
+	rootCert, err = k8s.GetCertificateFromSecret(c.providerNamespace, c.caBundleSecretName, rootCert, c.kubeClient)
 	if err != nil {
 		return nil, nil, errors.Errorf("Failed to synchronize certificate on Secrets API : %v", err)
+	}
+
+	if rootCert.GetPrivateKey() == nil {
+		return nil, nil, fmt.Errorf("Root cert does not have a private key: %w", certificate.ErrInvalidCertSecret)
 	}
 
 	tresorClient, err := tresor.New(
@@ -247,42 +184,6 @@ func (c *Config) getTresorOSMCertificateManager() (*certificate.Manager, debugge
 		return nil, nil, fmt.Errorf("error instantiating osm certificate.Manager for Tresor cert-manager : %w", err)
 	}
 	return tresorCertManager, tresorCertManager, nil
-}
-
-// GetCertFromKubernetes is a helper function that loads a certificate from a Kubernetes secret
-// The function returns an error only if a secret is found with invalid data.
-func GetCertFromKubernetes(ns string, secretName string, kubeClient kubernetes.Interface) (*certificate.Certificate, error) {
-	certSecret, err := kubeClient.CoreV1().Secrets(ns).Get(context.Background(), secretName, metav1.GetOptions{})
-	if err != nil {
-		// TODO(#3962): metric might not be scraped before process restart resulting from this error
-		log.Error().Str(errcode.Kind, errcode.GetErrCodeWithMetric(errcode.ErrFetchingCertSecret)).
-			Msgf("Could not retrieve certificate secret %q from namespace %q", secretName, ns)
-		return nil, errSecretNotFound
-	}
-
-	pemCert, ok := certSecret.Data[constants.KubernetesOpaqueSecretCAKey]
-	if !ok {
-		// TODO(#3962): metric might not be scraped before process restart resulting from this error
-		log.Error().Err(errInvalidCertSecret).Str(errcode.Kind, errcode.GetErrCodeWithMetric(errcode.ErrObtainingCertFromSecret)).
-			Msgf("Opaque k8s secret %s/%s does not have required field %q", ns, secretName, constants.KubernetesOpaqueSecretCAKey)
-		return nil, errInvalidCertSecret
-	}
-
-	pemKey, ok := certSecret.Data[constants.KubernetesOpaqueSecretRootPrivateKeyKey]
-	if !ok {
-		// TODO(#3962): metric might not be scraped before process restart resulting from this error
-		log.Error().Err(errInvalidCertSecret).Str(errcode.Kind, errcode.GetErrCodeWithMetric(errcode.ErrObtainingPrivateKeyFromSecret)).
-			Msgf("Opaque k8s secret %s/%s does not have required field %q", ns, secretName, constants.KubernetesOpaqueSecretRootPrivateKeyKey)
-		return nil, errInvalidCertSecret
-	}
-
-	cert, err := tresor.NewCertificateFromPEM(pemCert, pemKey)
-	if err != nil {
-		log.Error().Err(err).Msg("Failed to create new Certificate from PEM")
-		return nil, err
-	}
-
-	return cert, nil
 }
 
 // getHashiVaultOSMCertificateManager returns a certificate manager instance with Hashi Vault as the certificate provider
@@ -326,7 +227,7 @@ func (c *Config) getCertManagerOSMCertificateManager(options CertManagerOptions)
 		return nil, nil, fmt.Errorf("Opaque k8s secret %s/%s does not have required field %q", c.providerNamespace, c.caBundleSecretName, constants.KubernetesOpaqueSecretCAKey)
 	}
 
-	rootCert, err := certmanager.NewRootCertificateFromPEM(pemCert)
+	rootCert, err := certificate.NewFromPEM(pemCert, nil)
 	if err != nil {
 		return nil, nil, fmt.Errorf("Failed to decode cert-manager CA certificate from secret %s/%s: %s", c.providerNamespace, c.caBundleSecretName, err)
 	}
