+++ pkg/certificate/providers/config_test.go
@@ -3,12 +3,10 @@ package providers
 import (
 	"context"
 	"fmt"
-	"sync"
 	"testing"
 	"time"
 
 	"github.com/golang/mock/gomock"
-	"github.com/google/uuid"
 	tassert "github.com/stretchr/testify/assert"
 	corev1 "k8s.io/api/core/v1"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
@@ -16,15 +14,12 @@ import (
 	"k8s.io/client-go/rest"
 	"k8s.io/client-go/tools/clientcmd"
 
-	"github.com/openservicemesh/osm/pkg/certificate"
-	"github.com/openservicemesh/osm/pkg/certificate/providers/tresor"
 	"github.com/openservicemesh/osm/pkg/configurator"
 	"github.com/openservicemesh/osm/pkg/constants"
-	"github.com/openservicemesh/osm/pkg/tests"
 	"github.com/openservicemesh/osm/pkg/tests/certificates"
 )
 
-func TestNewCertificateProvider(t *testing.T) {
+func TestGenerateCertificateManager(t *testing.T) {
 	fakeClient := fake.NewSimpleClientset()
 	clientConfig := clientcmd.NewNonInteractiveDeferredLoadingClientConfig(
 		clientcmd.NewDefaultClientConfigLoadingRules(),
@@ -66,7 +61,7 @@ func TestNewCertificateProvider(t *testing.T) {
 	for i, tc := range testCases {
 		t.Run(fmt.Sprintf("Testing test case %d: %s", i, tc.name), func(t *testing.T) {
 			assert := tassert.New(t)
-			_, _, _, err := NewCertificateProvider(fakeClient, kubeConfig, mockConfigurator, tc.providerKind, "osm-system", "osm-ca-bundle", tc.tresorOpt, tc.vaultOpt, tc.certManagerOpt, nil)
+			_, _, _, err := GenerateCertificateManager(fakeClient, kubeConfig, mockConfigurator, tc.providerKind, "osm-system", "osm-ca-bundle", tc.tresorOpt, tc.vaultOpt, tc.certManagerOpt, nil)
 			assert.Equal(tc.expErr, err != nil)
 		})
 	}
@@ -130,119 +125,6 @@ func TestGetCertificateManager(t *testing.T) {
 				assert.NoError(err)
 			}
 		})
-	}
-}
-
-func TestSynchronizeCertificate(t *testing.T) {
-	assert := tassert.New(t)
-	kubeClient := fake.NewSimpleClientset()
-
-	// Create some cert, using tresor's api for simplicity
-	cert, err := tresor.NewCA("common-name", time.Hour, "test-country", "test-locality", "test-org")
-	assert.NoError(err)
-
-	wg := sync.WaitGroup{}
-	wg.Add(10)
-	certResults := make([]*certificate.Certificate, 10)
-
-	// Test synchronization, expect all routines end up with the same cert
-	for i := 0; i < 10; i++ {
-		go func(num int) {
-			defer wg.Done()
-
-			resCert, err := GetCertificateFromSecret("test", "test", cert, kubeClient)
-			assert.NoError(err)
-
-			certResults[num] = resCert
-		}(i)
-	}
-	wg.Wait()
-
-	// Verifiy all of them loaded the exact same cert
-	for i := 0; i < 9; i++ {
-		assert.Equal(certResults[i], certResults[i+1])
-	}
-}
-
-func TestGetCertificateFromKubernetes(t *testing.T) {
-	assert := tassert.New(t)
-
-	certPEM, err := tests.GetPEMCert()
-	assert.NoError(err)
-	keyPEM, err := tests.GetPEMPrivateKey()
-	assert.NoError(err)
-
-	ns := uuid.New().String()
-	secretName := uuid.New().String()
-
-	testCases := []struct {
-		secret       *corev1.Secret
-		expectError  bool
-		expectNilVal bool
-	}{
-		{
-			// Valid cert, valid test
-			secret: &corev1.Secret{
-				ObjectMeta: metav1.ObjectMeta{
-					Name:      secretName,
-					Namespace: ns,
-				},
-				Data: map[string][]byte{
-					constants.KubernetesOpaqueSecretCAKey:             certPEM,
-					constants.KubernetesOpaqueSecretRootPrivateKeyKey: keyPEM,
-				},
-			},
-			expectError:  false,
-			expectNilVal: false,
-		},
-		{
-			// Error when cert fetch is not present
-			secret:       nil,
-			expectError:  true,
-			expectNilVal: true,
-		},
-		{
-			// Error when CA key is missing
-			secret: &corev1.Secret{
-				ObjectMeta: metav1.ObjectMeta{
-					Name:      secretName,
-					Namespace: ns,
-				},
-				Data: map[string][]byte{
-					constants.KubernetesOpaqueSecretRootPrivateKeyKey: keyPEM,
-				},
-			},
-			expectError:  true,
-			expectNilVal: true,
-		},
-		{
-			// Error when Private Key is missing
-			secret: &corev1.Secret{
-				ObjectMeta: metav1.ObjectMeta{
-					Name:      secretName,
-					Namespace: ns,
-				},
-				Data: map[string][]byte{
-					constants.KubernetesOpaqueSecretCAKey: certPEM,
-				},
-			},
-			expectError:  true,
-			expectNilVal: true,
-		},
-	}
-
-	for _, testElement := range testCases {
-		kubeClient := fake.NewSimpleClientset()
-
-		if testElement.secret != nil {
-			_, err = kubeClient.CoreV1().Secrets(ns).Create(context.Background(), testElement.secret, metav1.CreateOptions{})
-			assert.NoError(err)
-		}
-
-		cert, err := GetCertFromKubernetes(ns, secretName, kubeClient)
-
-		assert.Equal(testElement.expectError, err != nil)
-		assert.Equal(testElement.expectNilVal, cert == nil)
 	}
 }
